before testing
