+++ src/openfermion/utils/_operator_utils.py
@@ -27,26 +27,6 @@ class OperatorUtilsError(Exception):
     pass
 
 
-def eigenspectrum(operator):
-    """Compute the eigenspectrum of an operator.
-
-    WARNING: This function has cubic runtime in dimension of
-        Hilbert space operator, which might be exponential.
-
-    Args:
-        operator: QubitOperator, InteractionOperator, FermionOperator,
-            InteractionTensor, or InteractionRDM.
-
-    Returns:
-        eigenspectrum: dense numpy array of floats giving eigenspectrum.
-    """
-    from openfermion.transforms import get_sparse_operator
-    from openfermion.utils import sparse_eigenspectrum
-    sparse_operator = get_sparse_operator(operator)
-    eigenspectrum = sparse_eigenspectrum(sparse_operator)
-    return eigenspectrum
-
-
 def count_qubits(operator):
     """Compute the minimum number of qubits on which operator acts.
 
@@ -89,20 +69,6 @@ def count_qubits(operator):
         raise TypeError('Operator of invalid type.')
 
 
-def is_identity(operator):
-    """Check whether QubitOperator of FermionOperator is identity.
-
-    Args:
-        operator: QubitOperator or FermionOperator.
-
-    Raises:
-        TypeError: Operator of invalid type.
-    """
-    if isinstance(operator, (QubitOperator, FermionOperator)):
-        return list(operator.terms) == [()]
-    raise TypeError('Operator of invalid type.')
-
-
 def commutator(operator_a, operator_b):
     """Compute the commutator of two operators.
 
@@ -121,39 +87,155 @@ def commutator(operator_a, operator_b):
     return result
 
 
-def save_operator(operator, file_name=None, data_directory=None):
-    """Save FermionOperator or QubitOperator to file.
+def eigenspectrum(operator):
+    """Compute the eigenspectrum of an operator.
+
+    WARNING: This function has cubic runtime in dimension of
+        Hilbert space operator, which might be exponential.
 
     Args:
-        operator: An instance of FermionOperator or QubitOperator.
+        operator: QubitOperator, InteractionOperator, FermionOperator,
+            InteractionTensor, or InteractionRDM.
+
+    Returns:
+        eigenspectrum: dense numpy array of floats giving eigenspectrum.
+    """
+    from openfermion.transforms import get_sparse_operator
+    from openfermion.utils import sparse_eigenspectrum
+    sparse_operator = get_sparse_operator(operator)
+    eigenspectrum = sparse_eigenspectrum(sparse_operator)
+    return eigenspectrum
+
+
+def is_identity(operator):
+    """Check whether QubitOperator of FermionOperator is identity.
+
+    Args:
+        operator: QubitOperator or FermionOperator.
+
+    Raises:
+        TypeError: Operator of invalid type.
+    """
+    if isinstance(operator, (QubitOperator, FermionOperator)):
+        return list(operator.terms) == [()]
+    raise TypeError('Operator of invalid type.')
+
+
+def _fourier_transform_helper(hamiltonian,
+                              grid,
+                              spinless,
+                              phase_factor,
+                              vec_func_1,
+                              vec_func_2):
+    hamiltonian_t = FermionOperator.zero()
+    normalize_factor = numpy.sqrt(1.0 / float(grid.num_points()))
+
+    for term in hamiltonian.terms:
+        transformed_term = FermionOperator.identity()
+        for ladder_op_mode, ladder_op_type in term:
+            indices_1 = grid_indices(ladder_op_mode, grid, spinless)
+            vec1 = vec_func_1(indices_1, grid)
+            new_basis = FermionOperator.zero()
+            for indices_2 in grid.all_points_indices():
+                vec2 = vec_func_2(indices_2, grid)
+                spin = None if spinless else ladder_op_mode % 2
+                orbital = orbital_id(grid, indices_2, spin)
+                exp_index = phase_factor * 1.0j * numpy.dot(vec1, vec2)
+                if ladder_op_type == 1:
+                    exp_index *= -1.0
+
+                element = FermionOperator(((orbital, ladder_op_type),),
+                                          numpy.exp(exp_index))
+                new_basis += element
+
+            new_basis *= normalize_factor
+            transformed_term *= new_basis
+
+        # Coefficient.
+        transformed_term *= hamiltonian.terms[term]
+
+        hamiltonian_t += transformed_term
+
+    return hamiltonian_t
+
+
+def fourier_transform(hamiltonian, grid, spinless):
+    """Apply Fourier transform to change hamiltonian in plane wave basis.
+
+    .. math::
+
+        c^\dagger_v = \sqrt{1/N} \sum_m {a^\dagger_m \exp(-i k_v r_m)}
+        c_v = \sqrt{1/N} \sum_m {a_m \exp(i k_v r_m)}
+
+    Args:
+        hamiltonian (FermionOperator): The hamiltonian in plane wave basis.
+        grid (Grid): The discretization to use.
+        spinless (bool): Whether to use the spinless model or not.
+
+    Returns:
+        FermionOperator: The fourier-transformed hamiltonian.
+    """
+    return _fourier_transform_helper(hamiltonian=hamiltonian,
+                                     grid=grid,
+                                     spinless=spinless,
+                                     phase_factor=+1,
+                                     vec_func_1=momentum_vector,
+                                     vec_func_2=position_vector)
+
+
+def get_file_path(file_name, data_directory):
+    """Compute file_path for the file that stores operator.
+
+    Args:
         file_name: The name of the saved file.
         data_directory: Optional data directory to change from default data
                         directory specified in config file.
 
+    Returns:
+        file_path (string): File path.
+
     Raises:
-        OperatorUtilsError: Not saved, file already exists.
-        TypeError: Operator of invalid type.
+        OperatorUtilsError: File name is not provided.
     """
-    file_path = get_file_path(file_name, data_directory)
-
-    if os.path.isfile(file_path):
-        raise OperatorUtilsError("Not saved, file already exists.")
-
-    if isinstance(operator, FermionOperator):
-        operator_type = "FermionOperator"
-    elif isinstance(operator, QubitOperator):
-        operator_type = "QubitOperator"
-    elif (isinstance(operator, InteractionOperator) or
-          isinstance(operator, InteractionRDM)):
-        raise NotImplementedError('Not yet implemented for InteractionOperator'
-                                  ' or InteractionRDM.')
+    if file_name:
+        if file_name[-5:] != '.data':
+            file_name = file_name + ".data"
     else:
-        raise TypeError('Operator of invalid type.')
+        raise OperatorUtilsError("File name is not provided.")
 
-    tm = operator.terms
-    with open(file_path, 'wb') as f:
-        marshal.dump((operator_type, dict(zip(tm.keys(),
-                                          map(complex, tm.values())))), f)
+    if data_directory is None:
+        file_path = DATA_DIRECTORY + '/' + file_name
+    else:
+        file_path = data_directory + '/' + file_name
+
+    return file_path
+
+
+def inverse_fourier_transform(hamiltonian, grid, spinless):
+    """Apply inverse Fourier transform to change hamiltonian in plane wave dual
+    basis.
+
+    .. math::
+
+        a^\dagger_v = \sqrt{1/N} \sum_m {c^\dagger_m \exp(i k_v r_m)}
+        a_v = \sqrt{1/N} \sum_m {c_m \exp(-i k_v r_m)}
+
+    Args:
+        hamiltonian (FermionOperator):
+            The hamiltonian in plane wave dual basis.
+        grid (Grid): The discretization to use.
+        spinless (bool): Whether to use the spinless model or not.
+
+    Returns:
+        FermionOperator: The inverse-fourier-transformed hamiltonian.
+    """
+
+    return _fourier_transform_helper(hamiltonian=hamiltonian,
+                                     grid=grid,
+                                     spinless=spinless,
+                                     phase_factor=-1,
+                                     vec_func_1=position_vector,
+                                     vec_func_2=momentum_vector)
 
 
 def load_operator(file_name=None, data_directory=None):
@@ -191,29 +273,36 @@ def load_operator(file_name=None, data_directory=None):
     return operator
 
 
-def get_file_path(file_name, data_directory):
-    """Compute file_path for the file that stores operator.
+def save_operator(operator, file_name=None, data_directory=None):
+    """Save FermionOperator or QubitOperator to file.
 
     Args:
+        operator: An instance of FermionOperator or QubitOperator.
         file_name: The name of the saved file.
         data_directory: Optional data directory to change from default data
                         directory specified in config file.
 
-    Returns:
-        file_path (string): File path.
-
     Raises:
-        OperatorUtilsError: File name is not provided.
+        OperatorUtilsError: Not saved, file already exists.
+        TypeError: Operator of invalid type.
     """
-    if file_name:
-        if file_name[-5:] != '.data':
-            file_name = file_name + ".data"
-    else:
-        raise OperatorUtilsError("File name is not provided.")
+    file_path = get_file_path(file_name, data_directory)
 
-    if data_directory is None:
-        file_path = DATA_DIRECTORY + '/' + file_name
-    else:
-        file_path = data_directory + '/' + file_name
+    if os.path.isfile(file_path):
+        raise OperatorUtilsError("Not saved, file already exists.")
 
-    return file_path
+    if isinstance(operator, FermionOperator):
+        operator_type = "FermionOperator"
+    elif isinstance(operator, QubitOperator):
+        operator_type = "QubitOperator"
+    elif (isinstance(operator, InteractionOperator) or
+          isinstance(operator, InteractionRDM)):
+        raise NotImplementedError('Not yet implemented for InteractionOperator'
+                                  ' or InteractionRDM.')
+    else:
+        raise TypeError('Operator of invalid type.')
+
+    tm = operator.terms
+    with open(file_path, 'wb') as f:
+        marshal.dump((operator_type, dict(zip(tm.keys(),
+                                          map(complex, tm.values())))), f)
