+++ src/openfermion/utils/_sparse_tools_test.py
@@ -19,10 +19,10 @@ import unittest
 from scipy.linalg import eigh, norm
 from scipy.sparse import csc_matrix
 
+from openfermion.hamiltonians import jellium_model, wigner_seitz_length_scale
 from openfermion.ops import FermionOperator, normal_ordered, number_operator
 from openfermion.transforms import get_sparse_operator, jordan_wigner
-from openfermion.utils import (fourier_transform, Grid, jellium_model,
-                               wigner_seitz_length_scale)
+from openfermion.utils import fourier_transform, Grid
 from openfermion.utils._jellium_hf_state import (
     lowest_single_particle_energy_states)
 from openfermion.utils._sparse_tools import *
