+++ src/openfermion/utils/_trotter_error.py
@@ -17,8 +17,8 @@ from math import sqrt, ceil
 from scipy.linalg import expm
 
 from openfermion.config import *
+from openfermion.hamiltonians import MolecularData
 from openfermion.ops import normal_ordered, QubitOperator
-from openfermion.utils import MolecularData
 
 
 def commutator(op1, op2):
